+++ bin/install-dependencies.sh
@@ -12,4 +12,4 @@ fi
 
 (mkdir -p $node_modules/flux-shutdown-handler-api && cd $node_modules/flux-shutdown-handler-api && wget -O - https://github.com/fluxfw/flux-shutdown-handler-api/archive/refs/tags/v2022-10-19-1.tar.gz | tar -xz --strip-components=1)
 
-(mkdir -p $node_modules/flux-studies-selfservice-frontend && cd $node_modules/flux-studies-selfservice-frontend && wget -O - https://github.com/fluxfw/flux-studies-selfservice-frontend/archive/refs/tags/v2022-10-21-1-alpha.tar.gz | tar -xz --strip-components=1 && npm ci --omit=dev)
+(mkdir -p $node_modules/flux-studies-selfservice-frontend && cd $node_modules/flux-studies-selfservice-frontend && wget -O - https://github.com/fluxfw/flux-studies-selfservice-frontend/archive/refs/tags/v2022-10-24-1-alpha.tar.gz | tar -xz --strip-components=1 && npm ci --omit=dev)