+++ src/Adapter/Data/PersonalData/MAX_HOUSE_NUMBER.mjs
@@ -1,1 +1,0 @@
-export const MAX_HOUSE_NUMBER = 9_999;