+++ src/Adapter/Data/PersonalData/MIN_HOUSE_NUMBER.mjs
@@ -1,1 +1,0 @@
-export const MIN_HOUSE_NUMBER = 1;